# Documentacion/openapi.yaml
openapi: 3.0.3
info:
  title: Plants App API
  description: |
    API REST del TFG "Plants App" — aplicación móvil de gestión de plantas con
    funcionalidades de comunidad, chat en tiempo real, recordatorios y datos climáticos.
    Arquitectura: Clean Architecture (Node.js + Express + MongoDB + Socket.IO).
  version: 1.0.0
  contact:
    name: TFG — 2 DAM
servers:
  - url: http://localhost:3000/api
    description: Servidor de desarrollo local

# ─────────────────────────────────────────────────────────────────────────────
# COMPONENTES COMUNES
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido tras login o registro.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        code:
          type: string
          example: UNAUTHORIZED
        message:
          type: string
          example: Token inválido o expirado.
      required: [status, code, message]

    # ── Paginación ────────────────────────────────────────────────────────────
...
    Plant:
      type: object
      properties:
        id:
          type: string
          example: 64f1a2b3c4d5e6f7a8b9c0d2
        userId:
          type: string
          example: 64f1a2b3c4d5e6f7a8b9c0d1
        name:
          type: string
          example: Monstera Deliciosa
        speciesId:
          type: string
          nullable: true
          description: ID de la especie del catálogo.
          example: 64f1a2b3c4d5e6f7a8b9c0e1
        photo:
          type: string
          nullable: true
          example: https://res.cloudinary.com/demo/image/upload/plant.jpg
        location:
          type: string
          nullable: true
          description: Ubicación del usuario propietario.
        plantLocation:
          type: string
          nullable: true
          description: Ubicación específica de la planta (ciudad).
        plantLocationLat:
          type: number
          format: double
          nullable: true
        plantLocationLon:
          type: number
          format: double
          nullable: true
        wateringFrequency:
          type: integer
          description: Frecuencia de riego en días.
          example: 7
        lightNeed:
          type: string
          enum: [low, medium, high]
          nullable: true
          description: Necesidad de luz.
        pruningFrequency:
          type: integer
          nullable: true
          description: Frecuencia de poda en días.
        notes:
          type: string
          nullable: true
        nextWatering:
          type: string
          format: date-time
          nullable: true
          example: '2026-03-07T08:00:00.000Z'
        nextPruning:
          type: string
          format: date-time
          nullable: true
        considerWeatherForWatering:
          type: boolean
          description: Si true, el cron de clima ajusta el riego según previsión.
        pendingRainAdjustment:
          type: object
          nullable: true
          description: |
            Reset de `nextWatering` pendiente de validación por el cron del
            día siguiente (decisión #169 / F7-043). Se rellena cuando
            `_processWeather` aplica un reset por previsión de lluvia y se
            limpia / aplica rollback en `_processYesterdayRain`.
          properties:
            resetAt:
              type: string
              format: date-time
              description: Momento en el que se aplicó el reset por previsión.
            previousNextWatering:
              type: string
              format: date-time
              nullable: true
              description: Valor de `nextWatering` ANTES del reset (null si no había).
            expectedMm:
              type: number
              description: Probabilidad/precipitación esperada que motivó el reset.
            locationLat:
              type: number
            locationLon:
              type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreatePlantRequest:
      type: object
      required: [name, speciesId, plantLocation, wateringFrequency, lightNeed, location]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: Monstera Deliciosa
        speciesId:
          type: string
          description: ID de la especie del catálogo (obligatorio — decisión #22).
        photo:
          type: string
          nullable: true
        location:
          type: string
          description: Ubicación del usuario propietario.
        plantLocation:
          type: string
          description: Ubicación específica de la planta (ciudad, obligatorio — decisión #61).
        plantLocationLat:
          type: number
          format: double
          nullable: true
        plantLocationLon:
          type: number
          format: double
          nullable: true
        wateringFrequency:
          type: integer
          minimum: 1
          example: 7
        lightNeed:
          type: string
          enum: [low, medium, high]
        pruningFrequency:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        considerWeatherForWatering:
          type: boolean
    UpdatePlantRequest:
      type: object
      properties:
        name:
          type: string
        speciesId:
          type: string
          nullable: true
        photo:
          type: string
          nullable: true
        location:
          type: string
        wateringFrequency:
          type: integer
          minimum: 1
        lightNeed:
          type: string
          enum: [low, medium, high]
        pruningFrequency:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        considerWeatherForWatering:
          type: boolean
        plantLocation:
          type: string
          nullable: true
        plantLocationLat:
          type: number
          format: double
          nullable: true
        plantLocationLon:
          type: number
          format: double
          nullable: true
        lastWatered:
          type: string
          format: date-time
          nullable: true
          description: Solo se envía internamente desde POST /plants/:id/water.

    # ── Comunidad ─────────────────────────────────────────────────────────────
...
    Notification:
      type: object
      properties:
        id:
          type: string
          example: 64f1a2b3c4d5e6f7a8b9c0d6
        userId:
          type: string
          example: 64f1a2b3c4d5e6f7a8b9c0d1
        type:
          type: string
          enum: [watering, pruning, harvest, fertilizing, repotting, custom, weather_rain, weather_storm, admin_warning]
          description: Tipo de notificación. weather_rain/weather_storm generadas por el cron de clima. admin_warning por acciones de moderación.
          example: watering
        message:
          type: string
          example: Es hora de regar tu Monstera Deliciosa.
        reminderId:
          type: string
          nullable: true
          description: Id del recordatorio asociado. Null para notificaciones de clima o cosecha.
        plantId:
          type: string
          nullable: true
          description: Id de la planta asociada. Null para notificaciones generales.
        isRead:
          type: boolean
          default: false
          example: false
        createdAt:
          type: string
          format: date-time

    # ── Reportes de incidencia ─────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: No autorizado — token ausente o inválido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Error de validación en el cuerpo de la petición.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
  - BearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /plants:
    get:
      tags: [Plants]
      summary: Listar plantas del usuario
      description: Devuelve todas las plantas que pertenecen al usuario autenticado.
      responses:
        '200':
          description: Lista de plantas.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Plant'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Plants]
      summary: Crear planta
      description: Crea una nueva planta para el usuario autenticado.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlantRequest'
      responses:
        '201':
          description: Planta creada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plant'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /plants/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: ID de la planta.
    get:
      tags: [Plants]
      summary: Obtener detalle de planta
      description: Devuelve los datos de una planta específica del usuario autenticado.
      responses:
        '200':
          description: Detalle de la planta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Plants]
      summary: Actualizar planta
      description: Actualiza los datos de una planta existente del usuario.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlantRequest'
      responses:
        '200':
          description: Planta actualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plant'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Plants]
      summary: Eliminar planta
      description: Elimina la planta del usuario. Soft-delete en la base de datos.
      responses:
        '204':
          description: Planta eliminada correctamente (sin contenido).
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /plants/{id}/water:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: ID de la planta.
    post:
      tags: [Plants]
      summary: Registrar riego manual
      description: |
        Registra que el usuario ha regado manualmente la planta.
        Actualiza `lastWatered` a la fecha actual y recalcula `nextWatering`
        sumando `wateringFrequency` días (con ajuste estacional si aplica).
      responses:
        '200':
          description: Riego registrado. Devuelve la planta actualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ════════════════════════════════════════════════════════════════════════════
  # COMMUNITY
  # ════════════════════════════════════════════════════════════════════════════
...
  /notifications:
    get:
      tags: [Notifications]
      summary: Obtener notificaciones del usuario
      description: Devuelve todas las notificaciones in-app del usuario autenticado, ordenadas por fecha descendente.
      responses:
        '200':
          description: Lista de notificaciones.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [Notifications]
      summary: Eliminar todas las notificaciones del usuario
      description: Elimina permanentemente todas las notificaciones del usuario autenticado.
      responses:
        '204':
          description: Notificaciones eliminadas.
        '401':
          $ref: '#/components/responses/Unauthorized'
  /notifications/read:
    put:
      tags: [Notifications]
      summary: Marcar todas las notificaciones como leídas
      description: Marca todas las notificaciones del usuario autenticado como leídas (isRead=true).
      responses:
        '204':
          description: Notificaciones marcadas como leídas.
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ════════════════════════════════════════════════════════════════════════════
  # REPORTS (incidencias de usuarios)
  # ════════════════════════════════════════════════════════════════════════════
...